# openapi.yaml
openapi: 3.1.0
info:
  title: DMV Heat Pumps API
  version: 0.1.0
  description: >
    API for sizing, savings, rebates, and optional lead capture for the DMV Heat Pumps site/widget.
servers:
  - url: https://example.com
    description: Replace with your deployment base URL
paths:
...
  /api/incentives:
    get:
      operationId: getIncentives
      summary: Query incentives by utility (and optionally state)
      parameters:
        - in: query
          name: utility
          required: true
          schema: { type: string }
        - in: query
          name: state
          required: false
          schema:
            type: string
            enum: ["DC", "MD", "VA"]
        - in: query
          name: active
          required: false
          description: "Set to false to include inactive programs."
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IncentivesResponse" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "500":
          description: Server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/report-requests:
    post:
      operationId: createReportRequest
      summary: Capture an email to send the full report later
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  stored: { type: boolean }
                  requestId: { type: string, nullable: true }
                  message: { type: string }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "500":
          description: Server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }

        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
components:
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
        details: {}
...
    IncentiveProgram:
      type: object
      required: [id, utility, programName, amountUsd]
      properties:
        id: { type: string }
        utility: { type: string }
        state:
          type: string
          enum: ["DC", "MD", "VA"]
          nullable: true
        programName: { type: string }
        amountUsd: { type: number }
        notes: { type: string, nullable: true }
        sourceUrl: { type: string, nullable: true }
        updatedAt: { type: string }
    IncentivesResponse:
      type: object
      required: [utility, activeOnly, totalEstimatedUsd, programs, disclaimer]
      properties:
        utility: { type: string }
        activeOnly: { type: boolean }
        totalEstimatedUsd: { type: number }
        programs:
          type: array
          items: { $ref: "#/components/schemas/IncentiveProgram" }
        disclaimer: { type: string }
    ReportRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        name: { type: string }
        zip: { type: string }
        context:
          type: object
          additionalProperties: true
...